Android Refactor (#4)
We did this refactor in order to get access to the native `Query`, which would allow us to easily implement Live Queries in the future. Even though we're not doing live queries at this time, this uses the native Query API instead of copying/rewriting the implementation for each update.